append architecture to docker tags
We had to revert the initial s390x container build in
kernel-patches/runner#4f09482052c1df07aa46586098ce83ace1c8bfe6

As the s390x container would be published, the arm64 and amd64
ones would be gone because the latest containers that would get built would update
tag `main` and that tag would now point to a container of a single
architecture (s390x).

This would be solved if we could build all arch with a single step that
targets all platforms, but due to the divergence of the s390x Dockerfile build
compare to archs natively supported by actions/runner, we cannot build s390x as
part of the same buildx batch and therefore cannot create a s390x container.

Instead, we will be using different tags for each arch which are keyed
on the architecture itself.
e.g we will have main-s390x, main-x86_64, main-aarch64 tags. Each runner
will pull the right tag based on their arch.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>