s390x: remove the volume
Volumes get created and deleted each time the container restart.
the non-s390x containers do not have this volume, so let's uniformize
this bit too.

We could benefit from volumes to keep the logs generated by the
actions-runner, but this would be done by using a consitent volume
across restarts (as opposed to new anonymous volume on each restarts).