[workflow][version_bump] Fix reviewers
As explained in #15, in order to be able to access team-reviewer list, automation
needs to have permission to access org level info (the team).
GITHUB_TOKENs are scope at repo level and do not have the permission to
fetch that info.
If we want this to work, we need a new GH App with the right permission.
For now, this is more work than benefit, let's just add our usernames
and review later if we see benefits for this type of auto-generated PR, we
can go ahead an provide a GH app wit the right permissions.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>