Exit with special status code in token.sh on error
Add a check for an empty response when requesting an access token on
runner startup. If no token was recieved, exit with status 199, which
then can be picked up by systemd instructing it to not restart the
runner container.

BPF CI runners went down recently due to the following problem:

  * For some runners, after successful authentication, on attempt to
    recieve a job the runner app gets "Runner not found" error, which
    appears to be an error returned from github broker service [1].

  * The systemd service maintaining the runner container automatically
    restarts it, leading to a frequent error -> restart loop.

  * Each time container starts, the token.sh script [2] requests a
    runner-specific github access token. However there is a 5k/hour
    limit for such requests [3].

  * Because there are dozens of runners we run out of tokens in
    minutes which effectively disables healthy runners too.

This token.sh customization will make sure that the runners affected
by this (or similar) issue don't also break healthy runners for more
than an hour.

[1] https://github.com/actions/runner/issues/3857
[2] https://github.com/myoung34/docker-github-actions-runner/blob/2.323.0/token.sh
[3] https://docs.github.com/en/rest/using-the-rest-api/rate-limits-for-the-rest-api?apiVersion=2022-11-28#primary-rate-limit-for-authenticated-users

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>