Cache libbpf/ci dependencies
BPF CI uses various actions from libbpf/ci repo. Many of them have
system package dependencies and usually action code installs what it
needs itself.

Our self-hosted runners are docker containers. We can take advantage
of the fact that we control the docker image and pre-install common
dependencies in the image, effectively caching them.

So far this was done by hardcoding a list of some packages in
dockerfiles. A better approach is to install exactly the same packages
as libbpf/ci actions.

The first attempt to do this [1] failed, because it turned out some
dependencies (such as iproute2 needed for some BPF selftest) were only
installed in the dockerfile, and not by a relevant action.
Additionally s390x image broke because of missing dumb-init for a
similar reason.

Now that libbpf/ci part of these issues is fixed [2], let's update the
dockerfiles.

[1] https://github.com/kernel-patches/runner/pull/78
[2] https://github.com/libbpf/ci/pull/186

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>